requirements.txt
awsebcli==3.10.6
blessed==1.14.2
botocore==1.7.1
cement==2.8.2
certifi==2017.7.27.1
Django==1.11.4
djangorestframework==3.6.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
jmespath==0.9.3
pathspec==0.5.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.9.1
semantic-version==2.5.0
tabulate==0.7.5
termcolor==1.1.0
websocket-client==0.44.0
